requirements.txt
asgiref==3.5.2
Django==3.2
django-environ==0.9.0
et-xmlfile==1.1.0
gunicorn==20.1.0
jpholiday==0.1.8
openpyxl==3.0.10
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
